Fetches data when the search button is pressed instead of when the user hits return after having entered a city name.
